chore(YALB-598): set header and footer globals on all example pages
